feat: Register dynamic proxy using Transaction Proxy FactoryBean implementation
Modified to allow dynamically created transaction proxy objects to be registered as beans in the Spring application context by implementing the FactoryBean interface.

-  Implemented TransactionProxyFactoryBean class
- Registered transaction proxy factory bean in the config configuration file
- Utilized transaction proxy bean in test code

Previously, proxies were automatically created but not registered in the application context. Now, the dynamic proxy object is registered and managed as a bean via the getObject() method.